Split out `Grace.Prompt.Types` module
… to help unwind an import cycle
